Refactored codeset in-mem store
Moved some things around to better define the in-mem codeset store
implementation as a standalone component:
* no local variables - the codeset in-mem store is instantiated at
runtime
* codeset store implementation details are no longer visible outside
the codeset_store.go file (i.e. this code can even be moved to its own
module). The external code only uses a generic `CodesetStore` interface
instance. This facilitates implementing other codeset back-ends in the
future without the need to change the consumer code.
* dependency injection is used to decouple the REST codeset service
from the back-end in-mem store implementation. The in-mem codeset store
object is instantiated in `main` and passed to the REST codeset service
as a generic `CodesetStore` interface instance.
* updated the signature of codeset store methods to accept a context.
This is a golang best practice and will come in handy if this or other
codeset store implementations need to deal with aspects concerning other
async operations (deadlines, cancellation signals etc.).