Refactored gitea gitadmin implementation
Made a few code changes to take full advantage of the dependency
inversion principle and further decouple the git admin abstraction
from the gitea implementation:

* changed the codeset store implementation so it uses a GitAdmin
interface instance instead of the concrete GiteaAdminClient
implementation. This completes the separation and removes the last
dependency that the codeset store had on the gitea back-end
implementation.
* used dependency injection to create a GiteaAdminClient in `main`
and provide it to the codeset store as a generic GitAdmin interface
instance. This other good side-effect of doing this is that the
fuseml-core server can now fail fast if the gitea client configuration
is missing or incorrect. In the context of running fuseml-core as a
k8s service, this will be reflected by a continuous fail-restart cycle
of the fuseml-core pod until the gitea server starts or until the
configuration is corrected.
* removed the intermediate gitadmin objects, as they were no longer
needed given the previous changes. For the lack of a better place,
the GitAdmin interface was moved with the rest of the codeset store
implementation.