Trigger pipeline run on workflow assign and add workflow interface
This commit creates a pipeline run when assigning a workflow to a
codeset. This pipeline run is created with the pipeline default values,
and if there is no default value for one of the parameters the pipeline
run will fail.

This commit also adds the interace "WorkflowBackend" with the functions
required to perform the FuseML workflow operations.