Refactored CLI using cobra/viper
This commit makes the following changes to the CLI code, to create
a customizable CLI framework that does not rely exclusively on goa
generated code:

* changes the way goa generates `gen/http/<service>/client/cli.go` files
belonging to the various services, to alter the signature of the payload
building functions to allow a wider range of data types to be passed
(i.e. bool, int, arrays), as opposed to allowing only JSON-encoded strings.
Given that cobra and pflags support these flag types, the
payload builder functions no longer need to perform decoding on
primitive types and on array types. This can be extended in the future
to also support map types where values are primitives, to allow simple
dictionaries to be supplied as command line args (e.g. `fuseml runnable
list --label mykey:myvalue --label mykey:myothervalue`).
* the second change made to the goa generated files is removing the
code in the main `gen/http/cli/../cli.go` file, which is now no longer
compatible with the `<service>/client/cli.go` changes. The main CLI code
is also modified to no longer rely on this generated code.
* remove GRPC code for the CLI
* implements a new CLI code structure based on cobra, viper and pflags. The
codeset commands and the root command are currently the only commands implemented
under the new CLI structure.
* moves the git client code under the new cli structure and changes the
codeset register command to use the git URL extracted from the codeset
instead of that supplied as the GITEA_URL env variable. This also
required some minor changes on the server side, to ensure the URL is
populated properly during registration (as well as re-registration,
which wasn't intended, but it's a nice side-effect that can be used
to push codeset updates by re-running `codeset register` again).