Enable passing the project password via env variable
When new codeset is created, FuseML creates a user for project and
returns the credentials back. It can still push the code using these
new credentials as it knows them.

However next time when user wants to register a codeset to the same
project, no new user is created, so FuseML needs to know which user and
password to use when pushing the code.

It can also happen that FuseML user changes the credentials by other
means (e.g. using Gitea interface) so there has to be a way to tell FuseML
about new password.