CLI: Fixed handling of env variables for global options
Due to a known limitation in the viper library [1], binding
cobra flags to a go variable and to a viper environment
variable at the same time is not supported.

This commit implements a workaround [2] that updates the flags
according to information extracted by viper from env variables
and configuration files after the cobra logic has finished loading
the flag variables.

The commit also adds support for using an optional fuseml CLI
configuration file where global configuration parameter values
can be stored. This is a yaml file that can be located either at
`$PWD/.fuseml/config.yaml` or at `$HOME/.fuseml/config.yaml`.

The priority in which configuration options are loaded is:
1. explicit set via command line args
2. environment variable
3. configuration file
4. default flag value

[1] https://github.com/spf13/viper/issues/671
[2] https://carolynvanslyck.com/blog/2020/08/sting-of-the-viper/