Use wire to initialize fuseml endpoints and its dependencies
The increased number of dependencies is creating a big block of initialization code
that's order-dependent, not very interesting, and at the same time it is hard to break
up this code cleanly, especially because some dependencies are used
multiple times.

Wire aim to simplify the management of initialization code by describing the
services and their dependencies, then Wire processes the resulting graph to figure
out ordering and how to pass each service what it needs. By Making changes to an
service dependency (changing its function signature or adding or removing an
initializer) Wire do the tedious work of generating initialization code for the
entire dependency graph.