fix: Unassign workflow when deleting assigned codeset
When a codeset is deleted and it has a workflow assigned to it, also
unassign the workflow from the codeset.

To perform that without creating a cyclic dependency between workflow
manager and codeset store, a pub/sub scheme was created. This allows
codeset store to notify workflow manager when a codeset that has
a workflow assigned to it is being deleted. With that notification
the workflow manager performs the unassign from the workflow to the
codeset being deleted.

Fixes: fuseml/fuseml#113