extensions: fix endpoint URL encoding
The URL endpoint attribute is used as a URL path token, which means
it needs to be URL-escaped by the CLI and URL-unescaped by the server.

Unfortunately, goa doesn't support this use-case, and worse, handling
the URL-escaping outside of goa clashes with the URL format
validation.

This commit removes the URL format validation for the endpoint URL
field and handles the URL escaping in the CLI/server code.